+++ lib/shard/src/operations/mod.rs
@@ -1,4 +1,4 @@
-pub mod operation_label;
+pub mod operation_name;
 pub mod optimization;
 pub mod payload_ops;
 pub mod point_ops;
